requirements.txt
coverage==4.0.1
Django==1.8.5
django-coverage==1.2.4
factory-boy==2.6.0
fake-factory==0.5.3
flake8==2.4.1
mccabe==0.3.1
pep8==1.5.7
pyflakes==0.8.1
wheel==0.24.0
